requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
mlflow>=2.8.0
dvc>=3.0.0
evidently>=0.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
kaggle>=1.5.16
python-dotenv>=1.0.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
black>=23.10.0
